+++ demo/classification/titanic/run_demo_titanic.sh
@@ -1,2 +1,5 @@
-# Use cross validation
+# Train model:
+#  -s 1   (use factorizarion machone for classification)
+#  -x acc (use accuracy metric)
+#  --cv   (use cross-validation)
 ../../xlearn_train ./titanic_train.txt -s 1 -x acc --cv