+++ doc/news.md
@@ -1,5 +1,12 @@
 ## News
 
+ - 08/03/2018 xLearn 0.3.0 version release. Main update:
+
+    * Fix bugs in previous version
+    * Solved the memory leak problem for on-disk learning
+    * Support TXT model checkpoint
+    * Support Scikit-Learn API
+
  - 18/12/2017 xLearn 0.2.0 version release. Main update:
 
     * Fix bugs in previous version